requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# Visualize_the_Board\Board.py: 10,11,12,13,14
# Visualize_the_Board\visualize_main.py: 2,3
kivy.deps.sdl2 == 0.1.18

# Reinforcement_Learning\Monte_Carlo_Search_Tree\MCTS_main.py: 1
# Reinforcement_Learning\Monte_Carlo_Search_Tree\deep_structure.py: 1
# Reinforcement_Learning\Monte_Carlo_Search_Tree\self_play.py: 1,4
# Reinforcement_Learning\game_state.py: 6
numpy == 1.14.5

# Reinforcement_Learning\Monte_Carlo_Search_Tree\MCTS_main.py: 3
# Reinforcement_Learning\Monte_Carlo_Search_Tree\self_play.py: 2
# Reinforcement_Learning\game_state.py: 4
# Visualize_the_Board\Board.py: 7
python_chess == 0.23.9

# Reinforcement_Learning\Monte_Carlo_Search_Tree\deep_structure.py: 4,5,6,7,8
torch == 0.1.2